+++ Source/OxyPlot.Tests/PlotModel/PlotModelTests.cs
@@ -90,7 +90,7 @@ namespace OxyPlot.Tests
         {
             var model = new PlotModel();
             var rc = Substitute.For<IRenderContext>();
-            ((IPlotModel)model).Render(rc, 0, 0);
+            ((IPlotModel)model).Render(rc, new OxyRect(0, 0, 0, 0));
         }
 
         /// <summary>
@@ -101,7 +101,7 @@ namespace OxyPlot.Tests
         {
             var model = new PlotModel { Padding = new OxyThickness(0) };
             var rc = Substitute.For<IRenderContext>();
-            ((IPlotModel)model).Render(rc, double.Epsilon, double.Epsilon);
+            ((IPlotModel)model).Render(rc, new OxyRect(0, 0, double.Epsilon, double.Epsilon));
         }
 
         /// <summary>