+++ Source/OxyPlot.WindowsForms/PngExporter.cs
@@ -94,7 +94,7 @@ namespace OxyPlot.WindowsForms
                 using (var rc = new GraphicsRenderContext(g) { RendersToScreen = false })
                 {
                     model.Update(true);
-                    model.Render(rc, this.Width, this.Height);
+                    model.Render(rc, new OxyRect(0, 0, this.Width, this.Height));
                 }
 
                 bm.SetResolution((float)this.Resolution, (float)this.Resolution);